requirements.txt
sounddevice
scipy
openai-whisper